+++ server/routes/auth.routes.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const authController = require("../controllers/auth.controller");
+
+const router = express.Router();
+
+router.post("/signup", authController.signup);
+router.post("/login", authController.login);
+module.exports = router;
